update to 1.4.4

+++ src/ground_vehicle.hpp
@@ -1,4 +1,4 @@
-/* $Id: ground_vehicle.hpp 24540 2012-09-21 20:31:06Z alberth $ */
+/* $Id: ground_vehicle.hpp 26714 2014-08-03 14:03:07Z frosch $ */
 
 /*
  * This file is part of OpenTTD.
@@ -113,9 +113,9 @@ struct GroundVehicle : public SpecializedVehicle<T, Type> {
 	 * Calculates the total slope resistance for this vehicle.
 	 * @return Slope resistance.
 	 */
-	inline int32 GetSlopeResistance() const
+	inline int64 GetSlopeResistance() const
 	{
-		int32 incl = 0;
+		int64 incl = 0;
 
 		for (const T *u = T::From(this); u != NULL; u = u->Next()) {
 			if (HasBit(u->gv_flags, GVF_GOINGUP_BIT)) {
